usb: dwc3: Correct datatype of base to unsigned long
Correct type of varibale base to unsigned long as
keeping it as int causes usb failures if MSB of
the base address is set.

+++ include/dwc3-uboot.h
@@ -13,7 +13,7 @@
 #include <linux/usb/otg.h>
 
 struct dwc3_device {
-	int base;
+	unsigned long base;
 	enum usb_dr_mode dr_mode;
 	u32 maximum_speed;
 	unsigned tx_fifo_resize:1;
